i128 hadd_x(Vec16c) experimental: undo the bias with a vector add
advantages: ILP (can overlap with the HILO64)

disadvantage: compilers are braindead and won't even reuse the same constant
as an input to _mm_xor and _mm_sub_epi16, even when it doesn't do
constant-propagation through _mm_slli_epi16.

Still, since compilers really want to load constants from memory,
instead of generating on the fly, loading two adjacent is fine.  They'll prob.
both be in the same cache line.

Still doesn't enable movd to memory,
because the result isn't sign-extended to 32bit.
pextrw sucks, and it's unlikely that we'd only be storing 16b anyway.