Vec4i operator *: use shift/and/or instead of 3 shuffles
This operation is probably often used inside of loops when it's used at
all.  The old version has 5 shuffles and 2 multiplies, so it will
bottleneck on shuffle throughput on almost all CPUs.  (Depending on
surrounding code).

It's probably worth introducing a constant so we can use vector ops that
run on more ports on more CPUs.

This will help a lot on SlowShuffle CPUs, and this is the only version
they can run.  It will also help a lot on Intel Haswell/Broadwell/Skylake.


Merom/Wolfdale, Bulldozer, Silvermont use different ports for pmul vs. psll.  Jaguar: mul on P0, shift on P01
Nehalem, SnB/IvB/HSW/BDW run pmuludq on the same port as psllq xmm,i, so that's not ideal

SKL runs pmul and shifts on p01, but shuffles only on p5 like Haswell, so
shuffle throughput is a major bottleneck for running the SSE2 version on SKL

So, this will help on new and old CPUs (except in cases where the constant
misses in cache and isn't).

It will help a LOT on SlowShuffle CPUs like Merom, which can't run the
SSE4.1 version.  This isn't the main motivation, though.  The motivation is
for the SSE2 version to run faster on lots of CPUs, including new ones,
for the benefit of software that doesn't do any dispatching and just uses
an SSE2 baseline.